chore: refactor TadoX to separate class
